fix: restored hamburger (see notes)
The 'hamburger menu' custom solution from the theme was removed by a previous commit

The feature has now been restored but the use of navbar-expand means that the hamburger menu icon will never show up - even on smaller widths (e.g. mobiles).

This is by design - at the moment it feels like the menu works well enough by wrapping to the next line at smaller sizes.

The hamburger/toggle (including the requisite bootstrap javascript code) has been left in for reference purposes as it was a little tricky to implement and we may have recourse to it in the future.