feat: move to bootstrap scss-style (see notes)
Swapping from importing the entire bootstrap css to importing only what is needed from the bootstrap scss plugin.

This should not change any of the styling, but does allow us to remove a significant amount of customisation and reimplementation of bootstrap css.

To maintain backwards compatibility with the theme's use of vars, we use the _boo

To reduce customisation and reimplementation of bootstraps's features, we use assets/scss/_bootstrap.scss as an entrypoint and map existing template vars to their bootstrap equivalents.