feat: restyled footer (see notes)
Removed the contact details (which we are not using) and tried to make the footer a little more compact and hopefully useful

We can still adjust the links in config.yaml to tweak what should be shown in the footer menu

Added data/footer.yml to be a datasource for social links and footer text. This could be edited by the CMS tools.