feat: standalone signup page (see notes)
The signup widget was in the footer previously but it was dragging down page load times without obvious benefit (being in the footer people are not automatically going to see it)

The signup form is now its own partial which is included in the index page as well as having its own /signup endpoint with consistent ui.

Could later introduce a shortcode or similar to allow content embedded signup forms.