fix: people layout issues (see notes)
I adjusted a few things on the people page and moved it to a custom layout with the leading text in the markdown file.

Not sure benefit of custom layout vs shortcode here but I didn't see we were likely to reuse the team page component
elsewhere so a shortcode didn't seem super necessary.